# feat: :sparkles: Add command to check the speed upload/download