--- a/.github/workflows/ci-build.yaml
+++ b/.github/workflows/ci-build.yaml
@@ -4,7 +4,7 @@ on:
     push:
         branches: ['*']
     pull_request:
-        branches: [master]
+        branches: [main]
 
 jobs:
     build:
@@ -12,7 +12,7 @@ jobs:
 
         strategy:
             matrix:
-                node-version: [16.x, 18.x, 19.x]
+                node-version: [16.x, 18.x, 19.x, 20.x]
 
         steps:
             - name: Check out code