--- a/package.json
+++ b/package.json
@@ -20,6 +20,7 @@
     "commander": "^10.0.1",
     "dotenv": "^16.0.3",
     "mime-types": "^2.1.35",
+    "pretty-bytes": "^6.1.0",
     "shelljs": "^0.8.5",
     "typescript": "^5.0.4"
   },