--- a/src/api/objects.ts
+++ b/src/api/objects.ts
@@ -4,12 +4,17 @@ import path from 'path';
 import { makeRequest } from './client';
 import { baseUrl, authToken } from '.';
 
-
 export async function fetchObject(containerName: string, objectName: string, outputPath?: string | undefined) {
     const url = `${baseUrl}/${containerName}/${objectName}`;
-    const response = await makeRequest('GET', url, {}, {
-        'X-Auth-Token': authToken,
-    }, 'stream');
+    const response = await makeRequest(
+        'GET',
+        url,
+        {},
+        {
+            'X-Auth-Token': authToken,
+        },
+        'stream',
+    );
 
     if (!outputPath) {
         outputPath = objectName;
@@ -28,34 +33,55 @@ export async function uploadObject(containerName: string, objectName: string, ov
     const filename = path.basename(objectName);
 
     const url = `${baseUrl}/${containerName}/${filename}`;
-    const response = await makeRequest('PUT', url, data, {
-        'X-Auth-Token': authToken,
-        'Content-Type': mimeType,
-    }, 'arraybuffer');
+    const response = await makeRequest(
+        'PUT',
+        url,
+        data,
+        {
+            'X-Auth-Token': authToken,
+            'Content-Type': mimeType,
+        },
+        'arraybuffer',
+    );
     return response.status === 201;
 }
 
 export async function fetchObjectList(containerName: string) {
     const url = `${baseUrl}/${containerName}`;
-    const response = await makeRequest('GET', url, {}, {
-        'X-Auth-Token': authToken,
-    });
+    const response = await makeRequest(
+        'GET',
+        url,
+        {},
+        {
+            'X-Auth-Token': authToken,
+        },
+    );
     return response.data;
 }
 
 export async function deleteObject(containerName: string, objectName: string) {
     const url = `${baseUrl}/${containerName}/${objectName}`;
-    const response = await makeRequest('DELETE', url, {}, {
-        'X-Auth-Token': authToken,
-    });
+    const response = await makeRequest(
+        'DELETE',
+        url,
+        {},
+        {
+            'X-Auth-Token': authToken,
+        },
+    );
     return response.status === 204;
 }
 
 export async function infoObject(containerName: string, objectName: string) {
     const url = `${baseUrl}/${containerName}/${objectName}`;
-    const response = await makeRequest('HEAD', url, {}, {
-        'X-Auth-Token': authToken,
-    });
+    const response = await makeRequest(
+        'HEAD',
+        url,
+        {},
+        {
+            'X-Auth-Token': authToken,
+        },
+    );
 
     if (response.status === 404) {
         return null;