--- a/src/index.ts
+++ b/src/index.ts
@@ -12,28 +12,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { program } from 'commander';
+import { Command, program } from 'commander';
 import chalk from 'chalk';
-import * as endpoints from "./api/endpoints";
-import * as config from "./utils/config";
-import * as speedtest from "./utils/speedtest";
+import * as endpoints from './api/endpoints';
+import * as config from './utils/config';
+import * as speedtest from './utils/speedtest';
+import { createFile, removeFile } from './utils/files';
 
 async function main() {
     // Set program details
-    program.name('magister')
-        .version('0.0.1')
-        .description('A CLI tool for interacting with On Machina™️');
+    program.name('magister').version('0.0.1').description('A CLI tool for interacting with On Machina™️');
 
     // Handle containers
-    program.command('container [containerName]')
+    program
+        .command('container [containerName]')
         .alias('c')
         .description('Create and switch to container')
-        .action(async (containerName) => {
+        .action(async containerName => {
             if (containerName === undefined) {
                 // If no container name is provided, use the default container
                 const configData = await config.getConfig();
                 containerName = configData.defaultContainer;
-                console.log(chalk.blue("Using container:", containerName));
+                console.log(chalk.blue('Using container:', containerName));
                 return;
             }
             // Check if the container already exists
@@ -48,33 +48,36 @@ async function main() {
             // Set the default container
             console.log(chalk.blue(`Using container: ${containerName}`));
             config.updateConfig({
-                defaultContainer: containerName
+                defaultContainer: containerName,
             });
         });
 
     // Upload a file
-    program.command('upload <file>')
+    program
+        .command('upload <file>')
         .alias('u')
         .alias('up')
         .description('Upload a file to the default container')
-        .action(async (objectName) => {
+        .action(async objectName => {
             const defaultContainer = (await config.getConfig()).defaultContainer;
             await endpoints.uploadObject(defaultContainer, objectName);
         });
 
     // Download a file
-    program.command('download <file>')
+    program
+        .command('download <file>')
         .alias('d')
         .alias('dl')
         .alias('down')
         .description('Download a file from the default container')
-        .action(async (objectName) => {
+        .action(async objectName => {
             const defaultContainer = (await config.getConfig()).defaultContainer;
             await endpoints.fetchObject(defaultContainer, objectName, objectName);
         });
 
     // List files
-    program.command('list')
+    program
+        .command('list')
         .alias('ls')
         .description('List objects in the default container')
         .action(async () => {
@@ -84,12 +87,13 @@ async function main() {
         });
 
     // Remove file
-    program.command('remove <file>')
+    program
+        .command('remove <file>')
         .alias('rm')
         .alias('del')
         .alias('delete')
         .description('Remove a file from the default container')
-        .action(async (objectName) => {
+        .action(async objectName => {
             const defaultContainer = (await config.getConfig()).defaultContainer;
             const deleted = await endpoints.deleteObject(defaultContainer, objectName);
 
@@ -102,12 +106,13 @@ async function main() {
         });
 
     // Get file info
-    program.command('info <file>')
+    program
+        .command('info <file>')
         .alias('i')
         .alias('inf')
         .alias('information')
         .description('Get information about a file from the default container')
-        .action(async (objectName) => {
+        .action(async objectName => {
             const defaultContainer = (await config.getConfig()).defaultContainer;
             const info = await endpoints.infoObject(defaultContainer, objectName);
             if (info === null) {
@@ -119,34 +124,40 @@ async function main() {
         });
 
     // Speed test
-    program.command('speedtest')
+    program
+        .command('speedtest')
         .alias('st')
         .description('Test upload and download speed')
-        .option('-s, --size <filesize>', 'file size to test with', '104857600')
-        .action(async () => {
+        .option('-s, --size <filesize>', 'file size in megabytes to test with', '10')
+        .action(async function (this: Command) {
+            const options = this.opts();
+
+            // Get default container
             const defaultContainer = (await config.getConfig()).defaultContainer;
 
-            // console.log(program.opts());
+            // Create test file
+            const fileSize: number = 1024 * 1024 * options.size;
+            console.log(chalk.blue(`Creating speed test file size: ${(fileSize / 1024 / 1024).toFixed(2)}MB`));
+            const testFile = 'speedtest.tmp';
+            await createFile(testFile, fileSize);
 
-            const fileSize: number = 1024 * 1024 * 10;
-            // const fileSize: number = program.opts().size as number;
-            // console.log(fileSize);
-
-            console.log(chalk.blue(`Testing speed with file size: ${(fileSize / 1024 / 1024).toFixed(2)}MB`));
-
+            // Upload test file
             console.log(chalk.blue('Testing upload speed...'));
-            const uploadStats = await speedtest.speedTestUpload(defaultContainer, fileSize);
-            console.log(chalk.blue('Upload complete!'));
-            console.log(chalk.blue(`Upload speed: ${uploadStats.speedPretty}`))
-            console.log(chalk.blue(`Time elapsed: ${uploadStats.timeElapsedPretty}`))
+            const uploadStats = await speedtest.speedTestUpload(defaultContainer, testFile, fileSize);
+            console.log(chalk.blue(`Upload speed: ${uploadStats.speedPretty}`));
+            console.log(chalk.blue(`Time elapsed: ${uploadStats.timeElapsedPretty}`));
 
+            // Download test file
             console.log(chalk.blue('Testing download speed...'));
-            const downloadStats = await speedtest.speedTestDownload(defaultContainer, fileSize);
-            console.log(chalk.blue('Download complete!'));
-            console.log(chalk.blue(`Download speed: ${downloadStats.speedPretty}`))
-            console.log(chalk.blue(`Time elapsed: ${downloadStats.timeElapsedPretty}`))
-        });
+            const downloadStats = await speedtest.speedTestDownload(defaultContainer, testFile, fileSize);
+            console.log(chalk.blue(`Download speed: ${downloadStats.speedPretty}`));
+            console.log(chalk.blue(`Time elapsed: ${downloadStats.timeElapsedPretty}`));
 
+            // Clean up
+            console.log(chalk.blue('Cleaning up...'));
+            await endpoints.deleteObject(defaultContainer, testFile);
+            await removeFile(testFile);
+        });
 
     // Parse commands
     program.parse();