--- a/src/utils/speedtest.ts
+++ b/src/utils/speedtest.ts
@@ -1,39 +1,34 @@
-import * as fs from 'fs/promises';
-import * as endpoints from "../api/endpoints";
-import { randomBytes } from 'node:crypto';
-// import * as prettyBytes from 'pretty-bytes';
+import * as endpoints from '../api/endpoints';
 let prettyBytes: any;
-
 import('pretty-bytes').then(pb => {
     prettyBytes = pb.default;
 });
 
 function formatDuration(milliseconds: number): string {
-    let seconds = Math.floor(milliseconds / 1000);  // convert to seconds
+    let seconds = Math.floor(milliseconds / 1000); // convert to seconds
 
     const hours = Math.floor(seconds / 3600);
     const minutes = Math.floor((seconds % 3600) / 60);
-    seconds = seconds % 60;  // remaining seconds
+    seconds = seconds % 60; // remaining seconds
 
-    let output = "";
+    let output = '';
 
     if (hours > 0) {
-        output += `${hours} hour${hours === 1 ? "" : "s"}`;
-        if (minutes > 0 || seconds > 0) output += ", ";
+        output += `${hours} hour${hours === 1 ? '' : 's'}`;
+        if (minutes > 0 || seconds > 0) output += ', ';
     }
 
     if (minutes > 0) {
-        output += `${minutes} minute${minutes === 1 ? "" : "s"}`;
-        if (seconds > 0) output += " and ";
+        output += `${minutes} minute${minutes === 1 ? '' : 's'}`;
+        if (seconds > 0) output += ' and ';
     }
 
     if (seconds > 0 || output.length === 0) {
-        output += `${seconds} second${seconds === 1 ? "" : "s"}`;
+        output += `${seconds} second${seconds === 1 ? '' : 's'}`;
     }
 
     return output;
 }
-
 
 type SpeedTestResult = {
     speed: number;
@@ -42,16 +37,12 @@ type SpeedTestResult = {
     timeElapsedPretty: string;
 };
 
-export async function speedTestUpload(defaultContainer: string, fileSize: number): Promise<SpeedTestResult> {
-    // Create test file
-    const testFile = 'speedtest.tmp';
-    await fs.writeFile(testFile, randomBytes(fileSize));
-
+export async function speedTestUpload(defaultContainer: string, fileName: string, fileSize: number): Promise<SpeedTestResult> {
     // Start timer
     const start = Date.now();
 
     // Upload file
-    await endpoints.uploadObject(defaultContainer, testFile);
+    await endpoints.uploadObject(defaultContainer, fileName);
 
     // End timer
     const end = Date.now();
@@ -59,9 +50,6 @@ export async function speedTestUpload(defaultContainer: string, fileSize: number
 
     // Calculate upload speed
     const speed = fileSize / (timeElapsed / 1000);
-
-    // Remove test file
-    await fs.unlink(testFile);
 
     return {
         speed: speed,
@@ -71,15 +59,12 @@ export async function speedTestUpload(defaultContainer: string, fileSize: number
     };
 }
 
-export async function speedTestDownload(defaultContainer: string, fileSize: number): Promise<SpeedTestResult> {
-    // Use test file
-    const testFile = 'speedtest.tmp';
-
+export async function speedTestDownload(defaultContainer: string, fileName: string, fileSize: number): Promise<SpeedTestResult> {
     // Start timer
     const start = Date.now();
 
     // Download file
-    await endpoints.fetchObject(defaultContainer, testFile);
+    await endpoints.fetchObject(defaultContainer, fileName);
 
     // End timer
     const end = Date.now();
@@ -87,9 +72,6 @@ export async function speedTestDownload(defaultContainer: string, fileSize: numb
 
     // Calculate upload speed
     const speed = fileSize / (timeElapsed / 1000);
-
-    // Remove test file
-    await fs.unlink(testFile);
 
     return {
         speed: speed,