Add a new command to test the upload and download speed.

```console
$ yarn start speedtest --size 100
yarn run v1.22.19
$ yarn run build && node dist/index.js speedtest --size 100
$ tsc
$ node build/post-build.js
Creating speed test file size: 100.00MB
Testing upload speed...
Upload speed: 11.8 MB/s
Time elapsed: 8 seconds
Testing download speed...
Download speed: 2.52 MB/s
Time elapsed: 41 seconds
Cleaning up...
Done in 54.16s.
```